Add VCR gem into project
Also configure VCR to delete authentication header. Move mocking
configuration to its own file.
