+++ spec/unit/cloudkeeper/aws/cloud_spec.rb
@@ -1,4 +1,5 @@
 require 'spec_helper'
+require 'mock_helper'
 
 module Cloudkeeper
   module Aws
